requirements.txt
tensorflow
numpy
opencv-python
pillow
streamlit
streamlit-drawable-canvas
